+++ messaging-core/src/main/java/com/nabla/sdk/messaging/core/data/message/ConversationContentRepositoryImpl.kt
@@ -106,6 +106,7 @@ internal class ConversationContentRepositoryImpl(
             is MessageInput.Media.Document -> Message.Media.Document(baseMessage, input.mediaSource)
             is MessageInput.Media.Image -> Message.Media.Image(baseMessage, input.mediaSource)
             is MessageInput.Text -> Message.Text(baseMessage, input.text)
+            is MessageInput.Media.Audio -> Message.Media.Audio(baseMessage, input.mediaSource)
         }
 
         return sendMessage(message)