+++ messaging-core/src/test/java/com/nabla/sdk/messaging/core/data/stubs/ConversationContentRepositoryStub.kt
@@ -93,6 +93,7 @@ internal class ConversationContentRepositoryStub(
             is MessageInput.Media.Document -> Message.Media.Document(baseMessage, input.mediaSource)
             is MessageInput.Media.Image -> Message.Media.Image(baseMessage, input.mediaSource)
             is MessageInput.Text -> Message.Text(baseMessage, input.text)
+            is MessageInput.Media.Audio -> Message.Media.Audio(baseMessage, input.mediaSource)
         }
 
         val conversationFlow = messagesFlowPerConversation[conversationId]!!