+++ CHANGELOG.md
@@ -10,7 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - `ConversationActivity` in conversations. A message is displayed when a provider joins a conversation.
 - `ConversationFragment` now displays system messages, with the right name and avatar.
-- Added support for receiving and playing voice messages. 
+- Added support for sending, receiving and playing voice messages.
 
 ### Changed
 - `MessageSender` is renamed to `MessageAuthor`.